rename v2 -> v3 and export in lib
