+++ src/lib.rs
@@ -1,4 +1,6 @@
 pub use siwe;
 
 pub mod v1;
-pub mod v2;
+pub mod v3;
+
+pub use v3::*;
